baseline/Conv-TasNet/pyproject.toml
[project]
name = "conv-tasnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lightning>=2.6.0",
    "pedalboard>=0.9.21",
    "python-dotenv>=1.2.1",
    "torchaudio>=2.9.1",
    "torchcodec>=0.9.1",
]
